backend/requirements.txt
fastapi
uvicorn[standard]
librosa
numpy
Pillow
